fix register controller bug

+++ src/utils/cloudinary.js
@@ -4,8 +4,8 @@ import fs from 'fs'
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-  api_key: process.env.CLOUD_API_KEY,
-  api_secret: process.env.CLOUD_API_SECRET,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
 const UploadOnCloudinary = async (localFilePath) =>{
@@ -15,11 +15,22 @@ const UploadOnCloudinary = async (localFilePath) =>{
        const response = await cloudinary.uploader.upload(localFilePath, {
            resource_type: "auto" 
         })
-        // file has been uploaded successfullly
-            console.log("file is uploaded on cloudinary",response.url)
+       
+            // console.log("file is uploaded on cloudinary",response)
+            
+            if (fs.existsSync(localFilePath)) {
+                     fs.unlinkSync(localFilePath);
+                }
             return response
     }catch(error){
-            fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        console.error("CLOUDINARY ERROR:", error.message);
+
+            if (fs.existsSync(localFilePath)) {
+                     fs.unlinkSync(localFilePath);
+                }
+
+    throw error; // 🔥 VERY IMPORTANT
+             // remove the locally saved temporary file as the upload operation got failed
     }
 }
